requirements.txt
pytest~=8.4.1
pytest-playwright
allure-pytest
playwright

allure-python-commons~=2.14.3
selenium~=4.34.2
webdriver-manager~=4.0.2